+++ src/features/user/components/index.ts
@@ -5,7 +5,6 @@ export { default as UserActivityStatItem } from "./UserActivityStatItem";
 // MyPage Components
 export { default as MyPageSidebarServer } from "./mypage/MyPageSidebar.server";
 export { default as MyPageTabContent } from "./mypage/MyPageTabContent";
-export { default as TabMenu } from "./mypage/MyPageTabMenu";
 
 // MyPage Tabs
 export { default as CommentsTab } from "./mypage/tabs/CommentsTab";
