--- src/features/user/components/mypage/MyPageTabMenu.tsx
@@ -1,39 +1,0 @@
-import { Tabs, TabsList, TabsTrigger } from "@/components/ui/tabs";
-import type { TabType } from "@/types/mypage";
-
-interface TabMenuProps {
-  tab: TabType;
-  onTabChange: (value: TabType) => void;
-  isOwner?: boolean;
-}
-
-const TAB_LABELS = {
-  posts: "작성 글",
-  likes: "좋아요",
-  comments: "작성 댓글",
-} as const;
-
-export default function MyPageTabMenu({ tab, onTabChange, isOwner = true }: TabMenuProps) {
-  const handleValueChange = (value: string) => {
-    onTabChange(value as TabType);
-  };
-
-  // 공개 프로필에서는 "작성 글" 탭만 표시
-  const visibleTabs = isOwner
-    ? Object.entries(TAB_LABELS)
-    : Object.entries(TAB_LABELS).filter(([key]) => key === "posts");
-
-  return (
-    <div className="w-full">
-      <Tabs value={tab} onValueChange={handleValueChange} className="w-full">
-        <TabsList className="flex w-full border-b border-[#F2F4F6] bg-transparent p-0">
-          {visibleTabs.map(([value, label]) => (
-            <TabsTrigger key={value} value={value} id={`${value}-tab`}>
-              {label}
-            </TabsTrigger>
-          ))}
-        </TabsList>
-      </Tabs>
-    </div>
-  );
-}
